+++ spine-unity/Assets/Spine/Editor/spine-unity/Editor/Utility/BuildSettings.cs
@@ -86,6 +86,7 @@ namespace Spine.Unity.Editor {
 		public const string SPINE_ALLOW_UNSAFE_CODE = "SPINE_ALLOW_UNSAFE";
 		public const string SPINE_AUTO_UPGRADE_COMPONENTS_OFF = "SPINE_AUTO_UPGRADE_COMPONENTS_OFF";
 		public const string SPINE_ENABLE_THREAD_PROFILING = "SPINE_ENABLE_THREAD_PROFILING";
+		public const string SPINE_DISABLE_LOAD_BALANCING = "SPINE_DISABLE_LOAD_BALANCING";
 
 		static bool IsInvalidGroup (BuildTargetGroup group) {
 			int gi = (int)group;